split i18n folder into multiple language files
- add "src/locale" folder
- make use of one translation file per language (eg. en.js)
- deprecate notice for "src/config/i18n" (that folder will be removed after v4)
