+++ src/locale/index.js
@@ -1,0 +1,19 @@
+import it from './it';
+import en from './en';
+import fi from './fi';
+import se from './se';
+import fr from './fr';
+import de from './de';
+import ro from './ro';
+
+const translations = {
+  it,
+  en,
+  fi,
+  se,
+  fr,
+  de,
+  ro,
+};
+
+export default translations;
